Remove code from crawler moved to itemmd
Use the ItemResponse rather than the Item type.
